+++ test/test_asgi.py
@@ -1,5 +1,5 @@
 import pytest
-from fastapi import Cookie
+from fastapi import Cookie, FastAPI
 from hypothesis import HealthCheck, given, settings
 
 import schemathesis
@@ -62,3 +62,33 @@ def test_not_app_with_asgi(schema):
         "Please, set `app` argument in the schema constructor or pass it to `call_asgi`",
     ):
         case.call_asgi()
+
+
+def test_base_url():
+    # See GH-1366
+    # When base URL has non-empty base path
+    raw_schema = {
+        "openapi": "3.0.3",
+        "info": {"version": "0.0.1", "title": "foo"},
+        "servers": [{"url": "https://example.org/v1"}],
+        "paths": {"/foo": {"get": {"responses": {"200": {"description": "OK"}}}}},
+    }
+
+    # And is used for an ASGI app
+    app = FastAPI()
+
+    @app.get("/v1/foo")
+    def read_root():
+        return {"Hello": "World"}
+
+    schema = schemathesis.from_dict(raw_schema)
+    strategy = schema["/foo"]["GET"].as_strategy()
+
+    @given(case=strategy)
+    @settings(max_examples=1, suppress_health_check=[HealthCheck.filter_too_much], deadline=None)
+    def test(case):
+        response = case.call_asgi(app)
+        # Then the base path should be respected and calls should not lead to 404
+        assert response.status_code == 200
+
+    test()